requirements.txt
beautifulsoup4==4.12.2
discord.py==2.3.2
numpy==1.25.2
matplotlib==3.7.2
plotly==5.16.1
openai==0.27.8
pandas==1.5.3
requests==2.31.0
scikit-learn==1.3.0
tqdm==4.66.1
